+++ app/src/main/java/com/android/tripbook/MainActivity.kt
@@ -1,0 +1,39 @@
+
+
+
+
+package com.android.tripbook
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.ui.Modifier
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            TripBookTheme {
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    PostCard(
+                        post = samplePost,
+                        onPostClick = { post ->
+                            // Handle post click - you can add navigation or other actions here
+                            println("Clicked on post: ${post.title}")
+                        }
+                    )
+                }
+            }
+        }
+    }
+}
+
+
